+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/query/type/EBPFProfilingTask.java
@@ -39,6 +39,7 @@ public class EBPFProfilingTask {
     private EBPFProfilingTargetType targetType;
     private long createTime;
     private long lastUpdateTime;
+    private String extensionConfigJson;
 
     /**
      * combine the same task