+++ docs/en/changes/changes.md
@@ -114,6 +114,7 @@
 * Fix not-time-series model blocking OAP boots in no-init mode.
 * Changed system variable `SW_SUPERDATASET_STORAGE_DAY_STEP` to `SW_STORAGE_ES_SUPER_DATASET_DAY_STEP` to be consistent with other ES storage related variables.
 * Fix ESEventQueryDAO missing metric_table boolQuery criteria.
+* Support dynamic config the sampling strategy in network profiling.
 
 #### UI